Core refactoring

+++ player/src/main/java/xyz/gianlu/librespot/player/codecs/VorbisCodec.java
@@ -207,7 +207,7 @@ public class VorbisCodec extends Codec {
             long granulepos = joggPacket.granulepos;
             if (granulepos != -1 && joggPacket.e_o_s == 0) {
                 granulepos -= samples;
-                granulepos -= BUFFER_SIZE * 6 * sampleSizeBytes(); // Account for buffer between the decoder and the player
+                granulepos -= (long) BUFFER_SIZE * 6 * sampleSizeBytes(); // Account for buffer between the decoder and the player
                 pcm_offset = granulepos;
             }
         }
